+++ benches/decimal128/bench_ln.mojo
@@ -9,9 +9,9 @@ from decimo.tests import (
     print_header,
     print_summary,
 )
-from python import Python, PythonObject
-from time import perf_counter_ns
-from collections import List
+from std.python import Python, PythonObject
+from std.time import perf_counter_ns
+from std.collections import List
 
 
 fn run_case(
@@ -37,14 +37,14 @@ fn run_case(
         var t0 = perf_counter_ns()
         for _ in range(iterations):
             _ = m_a.ln()
-        var tm = (perf_counter_ns() - t0) / iterations
+        var tm = (perf_counter_ns() - t0) / UInt(iterations)
         if tm == 0:
             tm = 1
 
         t0 = perf_counter_ns()
         for _ in range(iterations):
             _ = pa.ln()
-        var tp = (perf_counter_ns() - t0) / iterations
+        var tp = (perf_counter_ns() - t0) / UInt(iterations)
 
         var s = Float64(tp) / Float64(tm)
         sf.append(s)
