+++ benches/decimal128/bench_quantize.mojo
@@ -13,9 +13,9 @@ from decimo.tests import (
     print_header,
     print_summary,
 )
-from python import Python, PythonObject
-from time import perf_counter_ns
-from collections import List
+from std.python import Python, PythonObject
+from std.time import perf_counter_ns
+from std.collections import List
 
 
 fn get_mojo_rounding(mode_str: String) -> RoundingMode:
@@ -95,14 +95,14 @@ fn main() raises:
             var t0 = perf_counter_ns()
             for _ in range(iterations):
                 _ = m_a.quantize(m_quant, mojo_rm)
-            var tm = (perf_counter_ns() - t0) / iterations
+            var tm = (perf_counter_ns() - t0) / UInt(iterations)
             if tm == 0:
                 tm = 1
 
             t0 = perf_counter_ns()
             for _ in range(iterations):
                 _ = pa.quantize(py_quant, rounding=py_rm)
-            var tp = (perf_counter_ns() - t0) / iterations
+            var tp = (perf_counter_ns() - t0) / UInt(iterations)
 
             var s = Float64(tp) / Float64(tm)
             sf.append(s)
